pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


[tool.regarding]
release_name = ""
years = "2002-2021"

[tool.poetry]
name = "eyeD3"
version = "0.10a0"
description = "Python audio data toolkit (ID3 and MP3)"
authors = ["Travis Shirk <travis@pobox.com>"]
license = "GPL-3.0-or-later"
classifiers = [
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Sound/Audio :: Editors",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Operating System :: POSIX",
    "Natural Language :: English",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
keywords = ["id3", "mp3", "python"]
readme = "README.rst"
packages = [
    {include = "eyed3"},
]
include = [
    "poetry.lock",
    "README.rst", "LICENSE", "Makefile", "requirements.txt", "MANIFEST.in",
    "AUTHORS.rst", "CONTRIBUTING.rst", "HISTORY.rst",
    "tox.ini",
    "tests/**/*.py",
    "docs/",
    "requirements/*.txt",
    "examples/*",
    "eyed3/plugins/DisplayPattern.ebnf",
]
exclude = [
    "docs/_build",
]

[tool.poetry.scripts]
eyeD3 = "eyed3.main:_main"

[tool.poetry.dependencies]
python = "^3.6"
dataclasses = {version = "^0.8", python = "~3.6"}
filetype = "^1.0.7"
deprecation = "^2.1.0"
# yaml-plugin extra
"ruamel.yaml" = {version = "^0.16.12", optional = true}
# display-plugin extra
grako = {version = "^3.99.9", optional = true}
# art-plugin extra
Pillow = {version = "^8.0.1", optional = true}
pylast = {version = "^4.0.0", optional = true}
requests = {version = "^2.25.0", optional = true}
# Test extra
pytest = {version = "^6.2.1", optional = true}
coverage = {version = "^5.3.1", optional = true, extras = ["toml"]}
pytest-cov = {version = "^2.10.1", optional = true}
tox = {version = "^3.20.1", optional = true}
factory-boy = {version = "^3.1.0", optional = true}
flake8 = {version = "^3.8.4", optional = true}
check-manifest = {version = "^0.45", optional = true}

[tool.poetry.extras]
test = ["pytest", "pytest-cov", "tox", "factory-boy", "flake8",
        "check-manifest"]
yaml-plugin = ["ruamel.yaml"]
display-plugin = ["grako"]
art-plugin = ["Pillow", "pylast", "requests"]

[tool.poetry.dev-dependencies]
gitchangelog = {git = "https://github.com/nicfit/gitchangelog.git", rev = "nicfit.py"}
regarding = "^0.1.4"
wheel = "^0.36.2"
twine = "^3.3.0"
Sphinx = "^3.4.1"
sphinx_rtd_theme = "^0.5.0"
sphinx-issues = "^1.2.0"
cogapp = "^3.0.0"
paver = "^1.3.4"
"nicfit.py" = {version = "^0.8.7", extras = ["cookiecutter"]}

[tool.dephell]
[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}


[tool.coverage.html]
directory = "build/tests/coverage"
